requirements.txt
# AMI Simulation Platform - Dependencies
paho-mqtt>=2.1.0
influxdb-client>=1.38.0
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
